+++ examples/source-example/src/main/java/com/vance/source/MyAdapter.java
@@ -1,4 +1,4 @@
-package com.vance.samples;
+package com.vance.source;
 
 import com.linkall.vance.core.Adapter1;
 import io.cloudevents.CloudEvent;
